check status code

+++ stepper.py
@@ -105,8 +105,10 @@ for interval in IntervalTimer(10):
             img_data = f.read()
             img_name = str(current_time) + ".jpg"
         r = requests.post(url + "uploadRecycle", files={"file": img_data})
-        os.remove("/home/trashort/Pictures/" + str(current_time) + ".jpg")
-        vid.release()
+        if r.status_code == 201:
+            # turn off the camera
+            os.remove("/home/trashort/Pictures/" + str(current_time) + ".jpg")
+            vid.release()
     else:
         print("Organic waste")
         motor.enable(True)
@@ -118,5 +120,7 @@ for interval in IntervalTimer(10):
             img_data = f.read()
             img_name = str(current_time) + ".jpg"
         r = requests.post(url + "uploadOrganic", files={"file": img_data})
-        os.remove("/home/trashort/Pictures/" + str(current_time) + ".jpg")
-        vid.release()
+        if r.status_code == 201:
+            # turn off the camera
+            os.remove("/home/trashort/Pictures/" + str(current_time) + ".jpg")
+            vid.release()
